+++ packages/runtime-dom/src/index.ts
@@ -69,6 +69,7 @@ export const createApp = ((...args) => {
     container.innerHTML = ''
     const proxy = mount(container)
     container.removeAttribute('v-cloak')
+    container.setAttribute('data-v-app', '')
     return proxy
   }